Add test for dynamic_attribute example. Relates to #5
