+++ examples/dynamic_attribute/index.tsx
@@ -1,7 +1,8 @@
-import xs, { Stream } from 'xstream'
+import { Stream } from 'xstream'
 import { run, Drivers, FantasyObservable } from '@cycle/run'
 
-import { h, makeDOMDriver, DOMSource } from '../../src/cycle'
+import { makeDOMDriver, DOMSource } from '../../src/cycle'
+import { main } from './main'
 
 export interface Sources {
   DOM: DOMSource,
@@ -15,25 +16,6 @@ export interface Sinks {
 
 const drivers: Drivers<Sources, Sinks> = {
   DOM: makeDOMDriver('#app'),
-}
-
-function main({ DOM }: Sources): Sinks {
-  const buttonClick$ = DOM.selectEvents('.pressme', 'click')
-
-  const containerClass$ = buttonClick$
-    .fold(state => !state, false)
-    .map(state => state ? 'highlighted' : '')
-    .map(className => className + ' container')
-
-  const vtree = (
-    <div class$={containerClass$}>
-      <button class='pressme'>press me</button>
-    </div>
-  )
-
-  return {
-    DOM: xs.of(vtree),
-  }
 }
 
 run(main, drivers)
